test: Introduce AdkWebServerUITest
This WORKS and will not yet actually reproduce the
https://github.com/google/adk-java/issues/302 problem.

This because that is a packaging problem, which this
kind of test cannot detect. It still seems valuable to
have such a test here though anyway.

A future follow-up contribution on this might add the
missing test coverage for the actual packaging problem.