fix: Make `saveArtifact` in `CallbackContext` synchronous
The return type of the artifactService is Single, meaning it doesn't get scheduled to execute until it is subscribed. Today it's not subscribed. This is a bug leads to saveArtifact never executes.

Ideally the return type should be Rxjava instead of void, however since this is a open source library, changing return type is disruptive, the cl proposes to convert it as a blocking call, such that calling saveArtifact will be deterministically finished when it returns.

PiperOrigin-RevId: 817930873