Ensure all property types are translated when converting function calls to a2a
Python and Go implementations of ADK both copy function call and response
properties by copying the entire struct automatically. Java does this a property
at a time and so was missing the "id" property of the ADK event. Providing these
allows downstream agents to properly associate multiple tool calls and responses
to the same tool name.